add plan pl002 for patch upgrade
